+++ acceptance/main.tf
@@ -7,6 +7,7 @@ module "ansible_inv" {
   instances_aws = local.instances_aws
   instances_gcp = local.instances_gcp
   instances_azr = local.instances_azr
+  instances_vsp = local.instances_vsp
 }
 
 # validate generated inventory files with ad hoc ansible