+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Support direct input of AWS instance resource objects.
 - Support direct input of GCP instance resource objects.
 - Support direct input of Azure instance resource objects.
+- Support direct input of VSphere instance resource objects.
 
 ### 1.0.0
 - Initial Release